Restrict mjc resolver fix to physics import paths
