--- source/isaaclab/changelog.d/antoiner-mjc-frictionloss.skip
@@ -1,1 +1,0 @@
-Keeps the MJWarp ordering-resolver schema list in lockstep with the Newton manager import path.
